+++ src/common/enums/index.ts
@@ -11,4 +11,5 @@ export * from './user-type.enum';
 export * from './gender.enum';
 export * from './stat-method.enum';
 export * from './notification-method.enum';
-export * from './workflow-status.enum';
+export * from './order';
+export * from './suggest-status.enum';