+++ src/common/interfaces/index.ts
@@ -10,3 +10,5 @@ export * from './profile.interface';
 export * from './logistics';
 export * from './stat.interface';
 export * from './workflows';
+export * from './orders.interface';
+export * from './suggestions.interface';