+++ src/common/interfaces/notifications/notifier.interface.ts
@@ -1,12 +1,48 @@
-import { Resource } from '../../enums';
 import { Base } from '../base';
 
-//TODO: must be refactor after finishing tests.
-export interface Notifier<Meta = Record<string, unknown>> extends Base<Meta> {
-  resource: Resource;
+export interface KavenegarEntry {
+  messageid: number;
+
+  message: string;
+
+  status: number;
+
+  statustext: string;
+
+  sender: string;
+
   receptor: string;
-  template: string;
-  value_args?: string[];
-  seen?: Record<string, boolean>;
-  status?: Record<string, boolean>;
+
+  date: number;
+
+  cost: number;
 }
+
+export interface NotifierStatistics {
+  sms?: {
+    status: Record<string, unknown>; // or any other supported provider response type
+    message: string;
+  };
+  mail?: {
+    status: Record<string, unknown>;
+    message: string;
+  };
+  fcm?: {
+    status: Record<string, unknown>;
+    message: string;
+  };
+  socket?: {
+    status: Record<string, unknown>;
+    message: string;
+  };
+}
+
+export interface Notifier<Meta = Record<string, unknown>> extends Base<Meta> {
+  reason: string;
+  triggers: string;
+  value_args?: string[];
+  statistic?: NotifierStatistics;
+}
+
+export type NotificationNotifierInterface = Notifier;
+export type NotificationNotifierStatisticsInterface = NotifierStatistics;