+++ src/common/interfaces/workflows/state.interface.ts
@@ -1,8 +1,6 @@
-import { WorkflowStatus } from 'common/enums';
-
 export interface WorkflowState {
   ref: string;
   name?: string;
-  status: WorkflowStatus;
+  status: string;
   value?: unknown;
 }