+++ src/common/models/notifications/subs/index.ts
@@ -1,2 +1,3 @@
 export * from './options';
 export * from './configs';
+export * from './recipient-to.model';