+++ src/common/models/order/address.model.ts
@@ -1,0 +1,8 @@
+import { Geometry } from '../../types';
+
+export class AddressModel {
+  city!: string;
+  address?: string;
+  address_specification?: string;
+  geometry!: Geometry;
+}