+++ src/common/models/order/order-price.model.ts
@@ -1,0 +1,4 @@
+export class OrderPriceModel {
+  fix_cost?: boolean;
+  amount!: number;
+}